refactor: merge stripe-sync and stripe-sync-status files
